Transfer implementation of JsonResultSet and CsvResultSet to separate traits
